backend/requirements.txt
# IkerKeta pipeline library
ikerketa @ git+https://github.com/nkz-os/ikerketa.git@main

# FastAPI + ASGI
fastapi[standard]>=0.110.0
uvicorn[standard]>=0.27.0

# JWT validation
PyJWT[crypto]>=2.8.0

# HTTP client (used by ikerketa connectors)
httpx>=0.27.0

# Neo4j graph database
neo4j>=5.26.0

# Pydantic settings (env var / .env parsing)
pydantic-settings>=2.0.0

# YAML (seed data scripts)
PyYAML>=6.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0  # also used by TestClient
